+++ R/badge.R
@@ -19,7 +19,7 @@
 #'   - Experimental
 #'   - Maturing
 #'   - Stable
-#'   - Retired
+#'   - Superseded
 #'   - Archived
 #'   - Dormant
 #'   - Questioning
@@ -105,7 +105,7 @@ stages <- c(
   experimental = "orange",
   maturing = "blue",
   stable = "brightgreen",
-  retired = "orange",
+  superseded = "blue",
   archived = "red",
   dormant = "blue",
   questioning = "blue"
@@ -148,15 +148,5 @@ badge_start <- "<!-- badges: start -->"
 badge_end <- "<!-- badges: end -->"
 
 find_readme <- function() {
-  Rmd <- proj_path("README.Rmd")
-  if (file_exists(Rmd)) {
-    return(Rmd)
-  }
-
-  md <- proj_path("README.md")
-  if (file_exists(md)) {
-    return(md)
-  }
-
-  NULL
+  path_first_existing(proj_path(c("README.Rmd", "README.md")))
 }