+++ R/data.R
@@ -1,8 +1,8 @@
 #' Create package data
 #'
-#' `use_data()` makes it easy to save package data in the correct format.
-#' I recommend you save scripts that generate package data in `data-raw`:
-#' use `use_data_raw()` to set it up.
+#' `use_data()` makes it easy to save package data in the correct format. I
+#' recommend you save scripts that generate package data in `data-raw`: use
+#' `use_data_raw()` to set it up. You also need to document exported datasets.
 #'
 #' @param ... Unquoted names of existing objects to save.
 #' @param internal If `FALSE`, saves each object in its own `.rda`
@@ -53,6 +53,7 @@ use_data <- function(...,
   check_files_absent(proj_path(paths), overwrite = overwrite)
 
   ui_done("Saving {ui_value(unlist(objs))} to {ui_value(paths)}")
+  if (!internal) ui_todo("Document your data (see {ui_value('https://r-pkgs.org/data.html')})")
 
   envir <- parent.frame()
   mapply(