+++ R/logo.R
@@ -54,7 +54,7 @@ use_logo <- function(img, geometry = "240x278", retina = TRUE) {
   if (is.null(pd_link)) {
     ui_code_block("# {pkg} <img src={ui_path(logo_path)} align=\"right\" height=\"{height}\" />")
   } else {
-    ui_code_block("# {pkg} <a href={ui_path(pd_link, NA)}><img src={ui_path(logo_path)} align=\"right\" height=\"{height}\" /></a>")
+    ui_code_block("# {pkg} <a href={ui_value(pd_link)}><img src={ui_path(logo_path)} align=\"right\" height=\"{height}\" /></a>")
   }
 }