+++ R/package.R
@@ -62,16 +62,19 @@ use_remote <- function(package) {
 
 # Helpers -----------------------------------------------------------------
 
-## TO DO: make this less hard-wired to GitHub?
+## TO DO: make this less hard-wired to GitHub
 package_remote <- function(package) {
   desc <- desc::desc(package = package)
-  github_info <- desc$get(c("GithubUsername", "GithubRepo"))
+  # @jimhester and @gaborscardi say that the Github* fields are basically
+  # legacy, e.g., to support older tools like packrat
+  # Remote* fields are the way to go now
+  github_info <- desc$get(c("RemoteType", "RemoteUsername", "RemoteRepo"))
 
-  if (any(is.na(github_info))) {
+  if (!identical(github_info[["RemoteType"]], "github")) {
     ui_stop("{ui_value(package)} was not installed from GitHub.")
   }
 
-  glue_collapse(github_info, sep = "/")
+  glue::glue_data(as.list(github_info), "{RemoteUsername}/{RemoteRepo}")
 }
 
 refuse_package <- function(package, verboten) {