+++ tests/testthat/test-use-tidy.R
@@ -12,29 +12,6 @@ test_that("use_tidy_description() alphabetises dependencies", {
   expect_gt(grep("withr", desc), grep("gh", desc))
 })
 
-test_that("use_tidy_versions() specifies a version for dependencies", {
-  pkg <- scoped_temporary_package()
-  use_package("usethis")
-  use_package("desc")
-  use_package("withr", "Suggests")
-  use_package("gh", "Suggests")
-  use_tidy_versions()
-  desc <- readLines(proj_path("DESCRIPTION"))
-  desc <- grep("usethis|desc|withr|gh", desc, value = TRUE)
-  expect_true(all(grepl("\\(>= [0-9.]+\\)", desc)))
-})
-
-test_that("use_tidy_versions() does nothing for a base package", {
-  ## if we ever depend on a recommended package, could beef up this test a bit
-  pkg <- scoped_temporary_package()
-  use_package("tools")
-  use_package("stats", "Suggests")
-  use_tidy_versions()
-  desc <- readLines(proj_path("DESCRIPTION"))
-  desc <- grep("tools|stats", desc, value = TRUE)
-  expect_false(any(grepl("\\(>= [0-9.]+\\)", desc)))
-})
-
 test_that("use_tidy_eval() inserts the template file and Imports rlang", {
   skip_if_not_installed("roxygen2")
 
@@ -56,7 +33,7 @@ test_that("use_tidy_GITHUB-STUFF() adds and Rbuildignores files", {
       use_tidy_support()
       use_tidy_coc()
       expect_proj_file(".github/CONTRIBUTING.md")
-      expect_proj_file(".github/ISSUE_TEMPLATE.md")
+      expect_proj_file(".github/ISSUE_TEMPLATE/issue_template.md")
       expect_proj_file(".github/SUPPORT.md")
       expect_proj_file(".github/CODE_OF_CONDUCT.md")
       expect_true(is_build_ignored("^\\.github$"))
@@ -71,7 +48,7 @@ test_that("use_tidy_github() adds and Rbuildignores files", {
       scoped_temporary_package()
       use_tidy_github()
       expect_proj_file(".github/CONTRIBUTING.md")
-      expect_proj_file(".github/ISSUE_TEMPLATE.md")
+      expect_proj_file(".github/ISSUE_TEMPLATE/issue_template.md")
       expect_proj_file(".github/SUPPORT.md")
       expect_proj_file(".github/CODE_OF_CONDUCT.md")
       expect_true(is_build_ignored("^\\.github$"))